add condition to stop merge of non-ephimeral envs when PR is merged

+++ src/prune/main.go
@@ -292,17 +292,19 @@ func run(c *cli.Context) error {
 		log.Printf("Name: %s Duration: %s, Expired: %t, Merged: %t, Ephemeral: %t",
 			namespace.Name, time.Since(*namespace.UpdatedAt), expired, merged, namespace.IsEphemeral)
 
-		if expired || merged {
-			if !c.Bool("dryRun") {
-				if err := k8sClient.DeleteNamespace(namespace.Name); err != nil {
-					log.Printf("error deleting k8s namespace: %s", err)
-					continue
+		if merged && namespace.IsEphemeral {
+			continue
+			if expired || merged {
+				if !c.Bool("dryRun") {
+					if err := k8sClient.DeleteNamespace(namespace.Name); err != nil {
+						log.Printf("error deleting k8s namespace: %s", err)
+						continue
+					}
 				}
-			}
 
-			log.Printf("K8s namespace deleted: %s", namespace.Name)
+				log.Printf("K8s namespace deleted: %s", namespace.Name)
+			}
 		}
-	}
 	return nil
 }
